Switched to condiation csproj entries for net48 and net6
