+++ src/Prism.OptionalDependencies/DirectoryCatalog/SubDirectoryModuleCatalog.net48.cs
@@ -1,6 +1,4 @@
-﻿#if !NET5_0_OR_GREATER
-
-using Prism.Properties;
+﻿using Prism.Properties;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
@@ -255,5 +253,3 @@ namespace Prism.Modularity.OptionalDependencies
         }
     }
 }
-
-#endif
